+++ app/build.gradle.kts
@@ -8,7 +8,7 @@ android {
     compileSdk = 33
 
     defaultConfig {
-        applicationId = "template.app.id"
+        applicationId = "com.adammcneilly.compose.charts"
         minSdk = 21
         targetSdk = 33
         versionCode = 1
