+++ app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="template">
+    package="com.adammcneilly.compose.charts">
 
     <application
         android:allowBackup="true"
@@ -11,7 +11,7 @@
         android:theme="@style/Theme.AppTheme">
 
         <activity
-            android:name="template.MainActivity"
+            android:name="com.adammcneilly.compose.charts.MainActivity"
             android:exported="true"
             android:theme="@style/Theme.AppTheme.NoActionBar">
             <intent-filter>
