+++ app/src/main/java/com/adammcneilly/compose/charts/theme/Shape.kt
@@ -1,4 +1,4 @@
-package template.theme
+package com.adammcneilly.compose.charts.theme
 
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material3.Shapes
