+++ app/src/main/java/com/adammcneilly/compose/charts/theme/Theme.kt
@@ -1,4 +1,4 @@
-package template.theme
+package com.adammcneilly.compose.charts.theme
 
 import android.annotation.TargetApi
 import android.os.Build
@@ -23,7 +23,7 @@ private val lightColorScheme = lightColorScheme(
 
 @Composable
 @TargetApi(Build.VERSION_CODES.S)
-fun TemplateTheme(
+fun ChartsSampleTheme(
     darkTheme: Boolean = isSystemInDarkTheme(),
     dynamicTheme: Boolean = Build.VERSION.SDK_INT >= Build.VERSION_CODES.S,
     content: @Composable () -> Unit,
