+++ app/src/main/java/com/adammcneilly/compose/charts/theme/Type.kt
@@ -1,4 +1,4 @@
-package template.theme
+package com.adammcneilly.compose.charts.theme
 
 import androidx.compose.material3.Typography
 import androidx.compose.ui.text.TextStyle
